seed: make a shared seed system label validation helper
Address a UC20 TODO and move the seed system validation helper into a shared
package.
