+++ seed/validate.go
@@ -23,7 +23,6 @@ import (
 	"bytes"
 	"fmt"
 	"path/filepath"
-	"regexp"
 	"sort"
 
 	"github.com/snapcore/snapd/snap"
@@ -134,18 +133,5 @@ func ValidateFromYaml(seedYamlFile string) error {
 		return ve
 	}
 
-	return nil
-}
-
-// TODO:UC20: move these to internal, use also in seedwriter
-
-var validSeedSystemLabel = regexp.MustCompile("^[a-zA-Z0-9](?:-?[a-zA-Z0-9])+$")
-
-// validateSeedSystemLabel checks whether the string is a valid UC20 seed system
-// label.
-func validateUC20SeedSystemLabel(label string) error {
-	if !validSeedSystemLabel.MatchString(label) {
-		return fmt.Errorf("invalid seed system label: %q", label)
-	}
 	return nil
 }
